fix on SummaryDto, from long to BigDecimal

+++ src/test/java/com/fiserv/automation/api/dto/SummaryDto.java
@@ -1,14 +1,16 @@
 package com.fiserv.automation.api.dto;
 
+import java.math.BigDecimal;
+
 public class SummaryDto {
     public String estabelecimento;
     public String tipoSumarizacao;
     public String quantidade;
-    public long sumarizacao;
+    public BigDecimal sumarizacao;
     public String sumarizacaoConfirmado;
 
-    public long getSumarizacao() {
-        return sumarizacao;
+    public double getSumarizacao() {
+        return sumarizacao.doubleValue();
     }
 
     @Override
